+++ package.json
@@ -809,7 +809,7 @@
                 },
                 "appService.enableDetectors": {
                     "type": "boolean",
-                    "description": "Enables detectors that diagnose Linux web apps that diagnose container failures post deployment",
+                    "description": "Enables detectors that diagnose post-deployment Linux Container status",
                     "default": false
                 }
             }
