+++ src/commands/deploy/deploy.ts
@@ -74,7 +74,8 @@ export async function deploy(context: IActionContext, confirmDeployment: boolean
         }
     }
 
-    // cancel the previous detector check from the same web app
+    context.telemetry.properties.webAppSource = deployContext.webAppSource;
+
     if (newNodes.length > 0) {
         for (const newApp of newNodes) {
             if (newApp.fullId === node.fullId) {
@@ -156,7 +157,8 @@ export async function deploy(context: IActionContext, confirmDeployment: boolean
 
     await appservice.runPreDeployTask(deployContext, deployContext.deployFsPath, siteConfig.scmType, constants.extensionPrefix);
 
-    // moved to after prompts for telemetry purposes
+    // cancellation moved to after prompts while gathering telemetry
+    // cancel the previous detector check from the same web app
     const previousTokenSource: vscode.CancellationTokenSource | undefined = postDeployCancelTokens.get(node.id);
     if (previousTokenSource) {
         previousTokenSource.cancel();
@@ -165,8 +167,6 @@ export async function deploy(context: IActionContext, confirmDeployment: boolean
     await node.runWithTemporaryDescription("Deploying...", async () => {
         await appservice.deploy(nonNullValue(node).root.client, <string>deployContext.deployFsPath, deployContext, constants.showOutputChannelCommandId);
     });
-
-    context.telemetry.properties.webAppSource = deployContext.webAppSource;
 
     const tokenSource: vscode.CancellationTokenSource = new vscode.CancellationTokenSource();
     postDeployCancelTokens.set(node.id, tokenSource);
